requirements.txt
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
undetected-chromedriver==3.5.4 